+++ tests/Implementation/FailingLinkResolver.php
@@ -1,0 +1,30 @@
+<?php
+
+/**
+ * This file is part of the contentful/rich-text package.
+ *
+ * @copyright 2015-2018 Contentful GmbH
+ * @license   MIT
+ */
+
+declare(strict_types=1);
+
+namespace Contentful\Tests\RichText\Implementation;
+
+use Contentful\Core\Api\Link;
+use Contentful\Core\Api\LinkResolverInterface;
+use Contentful\Core\Resource\ResourceInterface;
+
+/**
+ * Basic implementation of a failing LinkResolverInterface for testing purposes.
+ */
+class FailingLinkResolver implements LinkResolverInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function resolveLink(Link $link, array $parameters = []): ResourceInterface
+    {
+        throw new \Exception();
+    }
+}