+++ tests/TestCase.php
@@ -15,63 +15,10 @@ use Contentful\RichText\Mark\MarkInterface;
 use Contentful\RichText\Node\NodeInterface;
 use Contentful\Tests\RichText\Implementation\Mark;
 use Contentful\Tests\RichText\Implementation\Node;
-use PHPUnit\Framework\TestCase as BaseTestCase;
-use function GuzzleHttp\json_decode as guzzle_json_decode;
+use Contentful\Tests\TestCase as BaseTestCase;
 
 class TestCase extends BaseTestCase
 {
-    /**
-     * Creates an empty assertion (true == true).
-     * This is done to mark tests that are expected to simply work (i.e. not throw exceptions).
-     * As PHPUnit does not provide convenience methods for marking a test as passed,
-     * we define one.
-     */
-    protected function markTestAsPassed()
-    {
-        $this->assertTrue(\true, 'Test case did not throw an exception and passed.');
-    }
-
-    /**
-     * @param string $file
-     * @param object $object
-     * @param string $message
-     */
-    protected function assertJsonFixtureEqualsJsonObject(string $file, $object, string $message = '')
-    {
-        $dir = $this->convertClassToFixturePath(\debug_backtrace()[1]['class']);
-        $this->assertJsonStringEqualsJsonFile($dir.'/'.$file, \GuzzleHttp\json_encode($object), $message);
-    }
-
-    /**
-     * @param string $file
-     * @param string $string
-     * @param string $message
-     */
-    protected function assertJsonFixtureEqualsJsonString(string $file, string $string, string $message = '')
-    {
-        $dir = $this->convertClassToFixturePath(\debug_backtrace()[1]['class']);
-        $this->assertJsonStringEqualsJsonFile($dir.'/'.$file, $string, $message);
-    }
-
-    /**
-     * @param string $file
-     *
-     * @return string
-     */
-    protected function getFixtureContent(string $file)
-    {
-        $dir = $this->convertClassToFixturePath(\debug_backtrace()[1]['class']);
-
-        return \file_get_contents($dir.'/'.$file);
-    }
-
-    protected function getParsedFixture(string $file)
-    {
-        $dir = $this->convertClassToFixturePath(\debug_backtrace()[1]['class']);
-
-        return guzzle_json_decode(\file_get_contents($dir.'/'.$file), \true);
-    }
-
     /**
      * This automatically determined where to store the fixture according to the test name.
      * For instance, it will convert a the class
@@ -82,7 +29,7 @@ class TestCase extends BaseTestCase
      *
      * @return string
      */
-    protected function convertClassToFixturePath(string $class)
+    protected function getClassFixturePath(string $class)
     {
         $class = \str_replace(__NAMESPACE__.'\\', '', $class);
         $class = \str_replace('\\', '/', $class);